+++ src/features/routes/reflection-list/full-view/FullViewReflectionPaperListArea.tsx
@@ -1,4 +1,3 @@
-import { useState } from "react";
 import Link from "next/link";
 import { format } from "date-fns";
 import { ja } from "date-fns/locale";
@@ -9,6 +8,7 @@ import type { ReflectionWithIncludeContent } from "@/src/api/reflection-api";
 import { ReflectionArticle } from "../../reflection-detail/article";
 import { Accordion } from "@/src/components/accordion";
 import { Button } from "@/src/components/button";
+import { useFullViewReflection } from "@/src/hooks/full-view/useFullViewReflection";
 import { theme } from "@/src/utils/theme";
 
 type FullViewReflectionPaperListAreaProps = {
@@ -17,51 +17,16 @@ type FullViewReflectionPaperListAreaProps = {
   userImage: string;
 };
 
-const groupReflectionsByMonth = (
-  reflections: ReflectionWithIncludeContent[]
-) => {
-  const groups = reflections.reduce(
-    (acc, reflection) => {
-      const month = format(new Date(reflection.createdAt), "yyyy-MM");
-      if (!acc[month]) {
-        acc[month] = [];
-      }
-      acc[month].push(reflection);
-      return acc;
-    },
-    {} as Record<string, ReflectionWithIncludeContent[]>
-  );
-
-  return Object.entries(groups).map(([month, items]) => ({
-    month,
-    reflections: items
-  }));
-};
-
-export const FullViewMonthlyReflectionListArea: React.FC<
+export const FullViewReflectionPaperListArea: React.FC<
   FullViewReflectionPaperListAreaProps
 > = ({ reflections, username, userImage }) => {
-  const [expandedMonths, setExpandedMonths] = useState<Record<string, boolean>>(
-    {}
-  );
-  const [isAscending, setIsAscending] = useState(false);
-
-  const toggleMonth = (month: string) => {
-    setExpandedMonths((prev) => ({
-      ...prev,
-      [month]: !prev[month]
-    }));
-  };
-
-  const toggleSortOrder = () => {
-    setIsAscending((prev) => !prev);
-  };
-
-  const groupedReflections = groupReflectionsByMonth(reflections);
-  const sortedReflections = [...groupedReflections].sort((a, b) => {
-    const comparison = a.month.localeCompare(b.month);
-    return isAscending ? comparison : -comparison;
-  });
+  const {
+    expandedMonths,
+    isAscending,
+    toggleMonth,
+    toggleSortOrder,
+    sortedReflections
+  } = useFullViewReflection(reflections);
 
   return (
     <Box mx={3}>
